+++ CHANGELOG.md
@@ -10,6 +10,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
+## [0.5.2] - 2023-10-13
+
+### Changed in 0.5.2
+
+- Changed from `int` to `int64` where required by the SenzingAPI
+- Update dependencies
+  - github.com/senzing/g2-sdk-go v0.7.3
+  - github.com/senzing/g2-sdk-go-base v0.3.1
+  - github.com/senzing/g2-sdk-proto/go v0.0.0-20231013142630-30a869751ff0
+  - google.golang.org/grpc v1.58.3
+
+### Deleted in 0.5.2
+
+- `g2product.ValidateLicenseFile`
+- `g2product.ValidateLicenseStringBase64`
+
 ## [0.5.1] - 2023-10-02
 
 ### Changed in 0.5.1